Changed UI, Added More colors.

+++ CarTuningConfigurator/Engine.cs
@@ -14,10 +14,10 @@ namespace CarTuningConfigurator
         public string? Type { get; set; }
         public string? Cylinder { get; set; }
         public int ImpactVelocity { get; set; }
-        public int ImpactAcceleration { get; set; }
+        public double ImpactAcceleration { get; set; }
         public int ImpactHorsePower { get; set; }
 
-        public Engine(string type, string cylinder, int impactVelocity, int impactAcceleration, int impactHorsePower)
+        public Engine(string type, string cylinder, int impactVelocity, double impactAcceleration, int impactHorsePower)
         {
             Type = type;
             Cylinder = cylinder;
@@ -28,7 +28,7 @@ namespace CarTuningConfigurator
 
         public Engine() { }
 
-        public Engine(int id, string name, string type, string cylinder, int level, double price, int impactVelocity, int impactAcceleration, int impactHorsePower)
+        public Engine(int id, string name, string type, string cylinder, int level, double price, int impactVelocity, double impactAcceleration, int impactHorsePower)
         {
             EngineId = id;
             Name = name;
